76: Move JacksonFormatter to the external package
