+++ app/src/test/groovy/gov/hhs/cdc/trustedintermediary/etor/order/OrderControllerTest.groovy
@@ -4,7 +4,7 @@ import gov.hhs.cdc.trustedintermediary.context.TestApplicationContext
 import gov.hhs.cdc.trustedintermediary.domainconnector.DomainRequest
 import gov.hhs.cdc.trustedintermediary.wrappers.Formatter
 import gov.hhs.cdc.trustedintermediary.wrappers.FormatterProcessingException
-import gov.hhs.cdc.trustedintermediary.wrappers.JacksonFormatter
+import gov.hhs.cdc.trustedintermediary.external.jackson.JacksonFormatter
 import spock.lang.Specification
 
 class OrderControllerTest extends Specification {
