+++ app/src/test/groovy/gov/hhs/cdc/trustedintermediary/etor/order/OrderTest.groovy
@@ -4,7 +4,7 @@ import gov.hhs.cdc.trustedintermediary.PojoTestUtils
 import gov.hhs.cdc.trustedintermediary.context.ApplicationContext
 import gov.hhs.cdc.trustedintermediary.context.TestApplicationContext
 import gov.hhs.cdc.trustedintermediary.wrappers.Formatter
-import gov.hhs.cdc.trustedintermediary.wrappers.JacksonFormatter
+import gov.hhs.cdc.trustedintermediary.external.jackson.JacksonFormatter
 import gov.hhs.cdc.trustedintermediary.wrappers.Logger
 import gov.hhs.cdc.trustedintermediary.wrappers.Slf4jLogger
 import spock.lang.Specification
